requirements.txt
django>=1.5
djangorestframework>=2.3
django-rest-swagger>=0.1.10
django-localflavor
coverage
South
